backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.3.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
PyMuPDF==1.23.26
langchain==0.1.9
langchain-groq==0.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 